+++ code/win32/win_syscon.c
@@ -1354,9 +1354,9 @@ HandleConsoleEvents(void)
   MSG msg;
 
   //pump the message loop
-  while(PeekMessage(&msg, NULL, 0, 0, PM_NOREMOVE))
+  while(PeekMessage(&msg, s_wcd.hWnd, 0, 0, PM_NOREMOVE))
   {
-    if (GetMessage(&msg, NULL, 0, 0) <= 0)
+    if (GetMessage(&msg, s_wcd.hWnd, 0, 0) <= 0)
     {
       Cmd_Clear();
       Com_Quit_f();